Use resourceManager instead of ResourceManager.getInstance()
We introduced this later, but in the meantime, you can use the
property resourceManager known from Components also in Studio Plugins.

The [ResourceBundle] annotation is like a dependency on a resource
bundle and should list at least all resource bundles used in the source
file. It might make sense to list more, like BlueprintDocumentTypes,
if you want to ensure a certain order of overwriting properties.

While at it, cleaned up the XML namespaces and the script block a bit.
In the script block, the import was no longer needed, an xtype is almost
always optional and mostly used for backwards-compatibility in CoreMedia
components, and the config field and the __initialize__ methods also
only have to be declared when they are actually used.
Only the native constructor unfortunately has to remain there.